backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
sentence-transformers
qdrant-client
rank-bm25
pypdf
